% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATAC_Seq_Tissue}
\alias{ATAC_Seq_Tissue}
\title{ATAC-Seq Significance for POL3 Active Genes}
\format{
A dataframe with 1415 rows and 19 variables:
\describe{
\item{Chr}{Coordinates for RNA of interest}
\item{Gene}{Chromosome where the RNA of interest is located}
\item{Index}{Unique RNA Sequence identifier - Can be retrieved from RNA Central}}
}
\source{
{Created by The VanBortle lab at UIUC to serve as an example}
}
\usage{
ATAC_Seq_Tissue
}
\description{
A subset of tissues with a list of POL3 Genes that are active.
}
\keyword{datasets}
