% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R
\name{circle}
\alias{circle}
\title{Entropy Plotter}
\usage{
circle(
  n,
  data,
  rect = F,
  back_alpha = 0.05,
  label = c("curve", "legend"),
  variables = colnames(data)
)
}
\arguments{
\item{n}{Total number of variables to account for Entropy calculation. It is the total number of columns}

\item{data}{A numerical dataframe. Each column represents the variable used to calculate the entropy and build the plot}

\item{rect}{A logical parameter. If TRUE, points will be displayed in a single rect line}

\item{back_alpha}{A numerical parameter, controls the intensity of the background for the plot}

\item{label}{It determines if the variables are labeled around the circle or displayed as a legend. Two options are available, 'Curve' or 'Legend'}

\item{variables}{A vector that contains the names of the columns that should appear in the plot when label == 'Curve'.}
}
\value{
Returns a list of objects. It contains the domination plot, a data frame with the calculated entropy, and a dataframe with the calculated categorical entropy
}
\description{
Creates a circular plot based on the Entropy for each row in a dataset.
}
\seealso{
[func(entropy)], [func(Qentropy)]
}
