% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_cpm.R
\name{cpm_normalization}
\alias{cpm_normalization}
\title{Counts Per Million}
\usage{
cpm_normalization(df, log_trans = FALSE)
}
\arguments{
\item{df}{A dataframe or matrix, must contain only numerical valeus}

\item{log_trans}{Logical attribute, determines if normalized data should be log2 transformed. Note: An extra count is added.}
}
\value{
A Matrix with normalized counts.
}
\description{
Normalizes a count matrix by the counts per million method
}
