% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations_entropy.R
\name{entropy}
\alias{entropy}
\title{Compute Shannon Entropy on row-normalized data}
\usage{
entropy(x, assay_name = NULL, new_assay_name = "Entropy")
}
\arguments{
\item{x}{A data.frame (with numeric columns)
or a SummarizedExperiment (with an assay of numeric data).}

\item{assay_name}{(SummarizedExperiment only) The name of the assay
to transform and compute Entropy on. If NULL, uses the first assay.}

\item{new_assay_name}{If you prefer to store Q-values in a
*new* assay, provide a name. By default `Entropy`}
}
\value{
\itemize{
  \item If \code{x} is a data.frame: returns the same data.frame in which
    numeric columns have been replaced by their row-wise proportions,
    and an \code{Entropy} column is appended.
  \item If \code{x} is a SummarizedExperiment: returns the same
    SummarizedExperiment in with a new assay (Default name is \code{Entropy})
    and \code{rowData(x)$Entropy} is added.
    }
}
\description{
Compute Shannon Entropy on row-normalized data
}
