% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_minmax.R
\name{minmax_normalization}
\alias{minmax_normalization}
\title{MinMax Normalization}
\usage{
minmax_normalization(x, min, max)
}
\arguments{
\item{x}{data that needed to be normalized}

\item{min}{the minimum value of the final normalized data}

\item{max}{the maximum value of the final normalized data}
}
\value{
a vector with normalized data
}
\description{
MinMax Normalization
}
