% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle.R
\name{plot_circle}
\alias{plot_circle}
\title{Dominance Plot in any Dimension - Plot Circle}
\usage{
plot_circle(
  n,
  data,
  str = FALSE,
  back_alpha = 0.05,
  label = c("curve", "legend"),
  variables = colnames(data),
  title = NULL,
  threshold = 0,
  col_variable = NULL,
  point_size = 3,
  text_size = 3,
  line_col = "gray90",
  out_line = "black",
  background_colors = NULL,
  point_fill_colors = NULL,
  point_line_colors = NULL
)
}
\arguments{
\item{n}{Total number of variables to account for entropy calculation.}

\item{data}{A numerical dataframe. Each column represent the variable used to calculate the entropy and build the plot}

\item{str}{A logical parameter. If TRUE points will be displayed in a single rect line}

\item{back_alpha}{A numerical parameter, controls the intensity of the background for the plot}

\item{label}{It determies if the variables are labeled around the circle or displayed as a legend. Two options are available 'curve' or 'legend'}

\item{variables}{A vector that contains the name of the columns that should appear in the plot when label == 'curve'}

\item{title}{Title of the plot}

\item{threshold}{A numerical value that determines the minmum values for rows to be filtered. Any row with a rowSums value less than this threshold is removed from the subsequent analysis. Default is 0}

\item{col_variable}{The name of a variable with categorical data that wants to be displayed. The function in general adds color to the observations based on their column names, if each observation has a category linked to it, by providing the name of that variable, data points will be colored accordingly}

\item{point_size}{Size for the plot points, default is 3}

\item{text_size}{Size for the text. It works when label is 'curve'}

\item{line_col}{Color for the plot outline.}

\item{out_line}{Color for the outtermost circle}

\item{background_colors}{Colors for the background of the plot}

\item{point_fill_colors}{Colors for the points in the plot}

\item{point_line_colors}{Colors for the outline of the point in the plot}
}
\value{
Returns a list of objects. It contains the circular dominance plot and a dataframe with the plotting coordinates, entropy and the dominant variable for each row
}
\description{
Creates a circular plot based on the entropy scores for each row in a dataset
}
\seealso{
[entropy()], [Qentropy()]
}
