% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R, R/plot_circle_frequency.R
\name{plot_circle_frequency}
\alias{plot_circle_frequency}
\title{Dominance Distribution Plotter}
\usage{
plot_circle_frequency(
  n,
  circle,
  single = FALSE,
  legend = TRUE,
  numb_columns = 1,
  filter_class = NULL,
  point_size = 2
)

plot_circle_frequency(
  n,
  circle,
  single = FALSE,
  legend = TRUE,
  numb_columns = 1,
  filter_class = NULL,
  point_size = 2
)
}
\arguments{
\item{n}{Number of Variables, must be the same as the ones used in the function}

\item{circle}{An object type list obtained from the plot_circle function}

\item{single}{A logical parameter. If TRUE a single plot will be produced otherwise a plot will be faceted}

\item{legend}{A logical parameter. If TRUE legends will appear}

\item{numb_columns}{A numerical value. Number of columns for plotting when single is FALSE}

\item{filter_class}{Name or names of classes of interest for display}

\item{point_size}{Size of the point}
}
\value{
Returns a list of objects, a frequency distribution plot for each category and the respective dataframe

Returns a list of objects, a frequency distribution plot for each category and the respective dataframe
}
\description{
This function depends on the plot.circle function when the parameter col_variable is used. It plots the dominance frequency distribution of each of the categories present in col_variable.

This function depends on the plot_circle function when the parameter col_variable is used. It plots the dominance frequency distribution of each of the categories present in col_variable.
}
\seealso{
[entropy()], [Qentropy()], [plot_circle()]
}
