% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rope.R
\name{plot_rope}
\alias{plot_rope}
\title{Dominance Plot in Two Dimensions - Plot Rope}
\usage{
plot_rope(
  data,
  push_text = 1,
  rope_width = 1,
  rope_color = "#CCCCCCCC",
  rope_border = T,
  col = c("red", "blue"),
  col_bg = "whitesmoke",
  pch = c(21, 21),
  pch_bg = 19,
  cex = 1,
  entropyrange = c(0, Inf),
  maxvaluerange = c(0, Inf),
  plotAll = T,
  label = T,
  title = "title",
  output_table = T
)
}
\arguments{
\item{data}{A data frame with two columns, one for each variable of interest}

\item{push_text}{The distance from the label to the plot}

\item{rope_width}{Width of the rope, default is 1}

\item{rope_color}{Color of the rope}

\item{rope_border}{A logical parameter, if TRUE, the rope will have a border}

\item{col}{A list of two colors used for the observations that pass the threshold}

\item{col_bg}{The color used for the observations that do not pass the threshold}

\item{pch}{A list of two shapes used for the observations that pass the threshold}

\item{pch_bg}{The shape of the points that do not pass the threshold}

\item{cex}{Point size}

\item{entropyrange}{Range to filter observations based on their entropy score}

\item{maxvaluerange}{Range to filter observations based on their value}

\item{plotAll}{A logical parameter, if TRUE, all the points will be plot, if FALSE only points above threshold}

\item{label}{A logical parameter. if TRUE, the names of each variable will be written on each side of the rope}

\item{title}{Title for the plot}

\item{output_table}{A logical parameter, if TRUE the dataframe used for making the plot will be returned}
}
\value{
A rope plot and a dataframe with the information used to build the plot
}
\description{
Creates a rope plot based on the entropy scores for each row in a dataset
}
\seealso{
[entropy()], [Qentropy()], [centmass()]
}
