% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_triangle.R
\name{plot_triangle}
\alias{plot_triangle}
\title{Plotting Triangle with center of mass}
\usage{
plot_triangle(
  data,
  entropyrange = c(0, 1),
  magnituderange = c(0, 1),
  col = c("red", "green", "blue"),
  output_table = T,
  plotAll = T,
  cex = 1,
  pch = 16
)
}
\arguments{
\item{data}{a table with three columns}

\item{entropyrange}{a vector with 2 elements in the range [0,1] i.e. c(minimum entropy , maximum entropy)
only the entries that has their entropy within this range are highlighted}

\item{magnituderange}{a vector with 2 elements in the range [0,1] i.e. c(minimum magnitude , maximum magnitude)
only the entries that has their magnitude within this range are highlighted}

\item{col}{a vector of 3 elements that contains color names}

\item{output_table}{a boolean value of T or F indicating whether to return the output table or not}

\item{plotAll}{a boolean value
If T, all the points that are out of entropyrange or magnitude range are plotted in whitesmoke color, else they are not plotted at all.}

\item{cex}{a numeric value indicating the size of the points}

\item{pch}{a integer value indicating different shapes of the points}
}
\value{
Returns a table with the coordinates of center of mass entropy(in the range of 0 and 1) and magnitude(in the range of 0 and 1)
}
\description{
Plotting Triangle with center of mass
}
