% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_quantnormalization.R
\name{quant_normalization}
\alias{quant_normalization}
\title{Quantile Normalization}
\usage{
quant_normalization(df, log_trans = F)
}
\arguments{
\item{df}{A dataframe or matrix, must contain only numerical values}

\item{log_trans}{Logical attribute, determines if normalized data should be log2 transformed. Note: An extra count is added.}
}
\value{
A Matrix with quantile normalized values
}
\description{
Normalizes a count matrix by the quantile normalization method
}
