% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnapol_chip}
\alias{rnapol_chip}
\title{RNA Polymerase CHIP-Seq Data}
\format{
A dataframe with 1635 rows and 4 variables:
\describe{
\item{UU}{Type of transcript}
\item{pol1}{Normalized peak call value for RNA Polymerase I Chip-Seq}
\item{pol2}{Normalized peak call value for RNA Polymerase II Chip-Seq}
\item{pol3}{Normalized peak call value for RNA Polymerase III Chip-Seq}}
}
\source{
{Created by The VanBortle lab at UIUC to serve as an example}
}
\usage{
rnapol_chip
}
\description{
A subset of data obtained from CHIP-Seq Data for RNA Polymerase I, II and III
}
\keyword{datasets}
