% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{rpkm}
\alias{rpkm}
\title{Reads Per Kilobase Million}
\usage{
rpkm(df, gene_length, log_trans = F)
}
\arguments{
\item{df}{A dataframe or matrix, must contain only numerical valeus}

\item{gene_length}{A vector with the length for the genes in the dataframe in basepair format}

\item{log_trans}{Logical attribute, determines if normalized data should be log2 transformed. Note: An extra count is added.}
}
\value{
A Matrix with normalized counts.
}
\description{
Normalizes a count matrix by the RPKM method
}
