% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizations_tpm.R
\name{tpm_normalization}
\alias{tpm_normalization}
\title{Transcripts per Million}
\usage{
tpm_normalization(df, gene_length, log_trans = F)
}
\arguments{
\item{df}{A dataframe or matrix, must contain only numerical valeus}

\item{gene_length}{A vector with the length for the genes in the dataframe in basepair format}

\item{log_trans}{Logical attribute, determines if normalized data should be log2 transformed. Note: An extra count is added.}
}
\value{
A Matrix with normalized counts.
}
\description{
Normalizes a count matrix by the TPM method
}
