% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTriangle.R
\name{vrtx}
\alias{vrtx}
\title{Vrtx - creates regular polygon}
\usage{
vrtx(n, plot = F)
}
\arguments{
\item{n}{an integer that indicates the number of vertex of the polygon}

\item{plot}{a boolean T/F that indicates wheteher to plot the polygon or not}
}
\value{
t
Returns the coordinates of the polygon
}
\description{
Returns coordinates of a regular polygon with n vertices. It plots the polygon with the flag "plot=T"
}
