% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations_centmass.R
\name{centmass}
\alias{centmass}
\title{Center of Mass}
\usage{
centmass(data, x = c(0, 1, 0.5), y = c(0, 0, sqrt(3)/2))
}
\arguments{
\item{data}{a table that provides mass of different objects}

\item{x}{a vector that provides x coordinates of  objects}

\item{y}{a vector that provides y coordinates of  objects}
}
\value{
Returns a table with 2 columns with x and y coordinate of the center of mass
}
\description{
Returns coordinates of center of mass from the data provided.
}
