% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R
\name{circle}
\alias{circle}
\title{Entropy Plotter}
\usage{
circle(
  n,
  data,
  rect = F,
  back_alpha = 0.05,
  label = c("curve", "legend"),
  labels = c(1:ncol(data))
)
}
\arguments{
\item{n}{Total number of variables to account for Entropy calculation. It is the total number of columns}

\item{data}{A numerical dataframe. Each column represents the variable used to calculate the entropy and build the plot}

\item{rect}{A logical parameter. If TRUE, points will be displayed in a single rect line}

\item{back_alpha}{A numerical parameter, controls the intensity of the background for the plot}

\item{label}{It determines if the variables are labeled around the circle or displayed as a legend. Two options are available, 'Curve' or 'Legend'}

\item{labels}{A vector that contains the numerical location of the variables that want to be labeled in the plot. The default is to present labels for each variable, but it can be modified if a vector with the numerical location of the variable of interest is provided}
}
\value{
Returns a list of objects. It contains the domination plot, a data frame with the calculated entropy, and a dataframe with the calculated categorical entropy
}
\description{
Entropy Plotter
}
\seealso{
[func(entropy), func(Qentropy)]
}
