% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R
\name{circle}
\alias{circle}
\title{Entropy Plotter}
\usage{
circle(
  n,
  data,
  str = F,
  back_alpha = 0.05,
  label = c("curve", "legend"),
  variables = colnames(data),
  title = NULL,
  threshold = 0,
  col_variable = NULL
)
}
\arguments{
\item{n}{Total number of variables to account for Entropy calculation. It is the total number of columns}

\item{data}{A numerical dataframe. Each column represents the variable used to calculate the entropy and build the plot}

\item{str}{A logical parameter. If TRUE, points will be displayed in a single rect line}

\item{back_alpha}{A numerical parameter, controls the intensity of the background for the plot}

\item{label}{It determines if the variables are labeled around the circle or displayed as a legend. Two options are available, 'Curve' or 'Legend'}

\item{variables}{A vector that contains the names of the columns that should appear in the plot when label == 'Curve'.}

\item{title}{Title of the plot.}

\item{threshold}{A numerical value that determines the minimum value for rows to be filtered. Any row with a rowSums value less than this threshold is removed from the subsequent analysis. Default is 0}

\item{col_variable}{The name of a variable with categorical data that wants to be displayed. The function in general adds color to the observations based on their column names, if each observation has a categorical value linked to it, by providing the name of that variable, data points will be colored accordingly.}
}
\value{
Returns a list of objects. It contains the domination plot, an a dataframe with the plotting coordinates with entropy and dominant variable for each point
}
\description{
Creates a circular plot based on the Entropy for each row in a dataset.
}
\seealso{
[entropy()], [Qentropy()]
}
