% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R
\name{entropy}
\alias{entropy}
\title{Entropy - Shannon Information}
\usage{
entropy(test)
}
\arguments{
\item{test}{A Data frame, must contain numerical columns}
}
\value{
Returns an extra column to the dataframe with a value related to the Entropy value. :)
}
\description{
Calculates the Entropy value for each row across numerical columns.
}
\examples{
data(rnapol_score)
test =  entropy(rnapol_score)



}
\references{
https://en.wikipedia.org/wiki/Entropy_(information_theory)
}
