% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_plot.R
\name{plot_circle_frequency}
\alias{plot_circle_frequency}
\title{Dominance Distribution Plotter}
\usage{
plot_circle_frequency(n, circle)
}
\arguments{
\item{n}{Number of Variables, must be the same as the ones used in the function}

\item{circle}{An object type list obtained from the plot.circle function}
}
\value{
Returns a list of objects, a frequency distribution plot for each category and the respective dataframe
}
\description{
This function depends on the plot.circle function when the parameter col_variable is used. It plots the dominance frequency distribution of each of the categories present in col_variable.
}
