% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations_Qentropy.R, R/entropy_plot.R
\name{Qentropy}
\alias{Qentropy}
\title{Categorical Entropy}
\usage{
Qentropy(test)

Qentropy(test)
}
\arguments{
\item{test}{A Data frame, must contain numerical columns and Entropy should be calculated already}
}
\value{
Returns the data frame with modified values for numerical columns.

Returns the data frame with modified values for numerical columns.
}
\description{
Calculates the Categorical Entropy value related to each column. Entropy calculation should be performed before using this function

Calculates the Categorical Entropy value related to each column. Entropy calculation should be performed before using this function
}
\examples{
data(rnapol_score)
test =  entropy(rnapol_score)
test = Qentropy(test)

data(rnapol_score)
test =  entropy(rnapol_score)
test = Qentropy(test)

}
\seealso{
[entropy()]

[entropy()]
}
